+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # intellij-jenv-plugin Changelog
 
 ## [Unreleased]
+- adjust the position of the jEnv status bar 
+- remove duplicate jEnv jdk. if jEnv jdk canonical path exists, delete all. if jEnv jdk home path has one more, delete them until one exists.
 
 ## [0.0.4] - 2023-12-02